requirements.txt
Flask==3.1.0
holidays==0.67
ipython==8.12.3
matplotlib==3.6.3
pandas_ta==0.3.14b0
prettytable==3.14.0
PyMySQL==1.1.1
pytz==2024.1
QuantLib==1.37
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.2
ta==0.11.0
tabulate==0.8.10
termcolor==1.1.0
yfinance==0.2.54
